requirements.txt
python-telegram-bot==20.7
groq>=0.9.0